+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueModel.kt
@@ -6,15 +6,13 @@ import org.simple.clinic.facility.Facility
 
 @Parcelize
 data class OverdueModel(
-    val facility: Facility?,
-    val overdueAppointmentOlds: List<OverdueAppointment_Old>?
+    val facility: Facility?
 ) : Parcelable {
 
   companion object {
     fun create(): OverdueModel {
       return OverdueModel(
-          facility = null,
-          overdueAppointmentOlds = null
+          facility = null
       )
     }
   }
@@ -22,17 +20,10 @@ data class OverdueModel(
   val isDiabetesManagementEnabled: Boolean
     get() = facility!!.config.diabetesManagementEnabled
 
-  val hasLoadedOverdueAppointments: Boolean
-    get() = overdueAppointmentOlds != null
-
   val hasLoadedCurrentFacility: Boolean
     get() = facility != null
 
   fun currentFacilityLoaded(facility: Facility): OverdueModel {
     return copy(facility = facility)
-  }
-
-  fun overdueAppointmentsLoaded(appointmentOlds: List<OverdueAppointment_Old>): OverdueModel {
-    return copy(overdueAppointmentOlds = appointmentOlds)
   }
 }